add update invoice in odoo

+++ app/Http/Controllers/admin/AdminPaymentAttemptController.php
@@ -233,14 +233,21 @@ class AdminPaymentAttemptController extends Controller
             return redirect()->back()->with('alert-danger', $msg);
         }
 
+        $result = $this->deletePaymentInOdoo(["payment_code_abnai" => $attempt->id]);
+
+        if(!$result['status']){
+            return redirect()->back()
+                ->with('alert-danger', $result['message']);
+        }
+
         if ($attempt->delete()) {
             $transaction->update_transaction($attempt);
 
             return redirect()->route('students.contracts.transactions.index', [$student, $contract, $transaction])
-                ->with('alert-success', 'تم حضف الدقغة  ينجاح');
+                ->with('alert-success', 'تم حذف الدقغة ينجاح');
         }
         return redirect()->back()
-            ->with('alert-danger', 'فشل حضف الدقغة  ');
+            ->with('alert-danger', 'فشل حذف الدقغة  ');
     }
 
     public function confirmPaymentAttempt($PaymentAttempt, $transaction, $request)
